backend/requirements.txt
fastapi==0.115.0
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.6.1
email-validator==2.1.0
sendgrid==6.10.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.27
python-dotenv==1.0.1
requests==2.31.0 